+++ src/js/menu.js
@@ -31,9 +31,29 @@ function refreshData() {
   toggleMenu();
 }
 
+async function toFixtures() {
+  const currentPage = await getCurrentPage();
+  chrome.storage.local.set({ previousPage: currentPage });
+  showPage('fixtures-overview');
+  toggleMenuStyle();
+}
+
+async function toStatistics() {
+  const currentPage = await getCurrentPage();
+  chrome.storage.local.set({ previousPage: currentPage });
+  showPage('statistics-overview');
+  toggleMenuStyle();
+}
+
 document.addEventListener('DOMContentLoaded', () => {
   const menuIcon = document.getElementById('fpl-menu');
   menuIcon.addEventListener('click', toggleMenu);
+
+  const fixtures = document.getElementById('fixtures');
+  fixtures.addEventListener('click', toFixtures);
+
+  const statistics = document.getElementById('statistics');
+  statistics.addEventListener('click', toStatistics);
 
   const logoutButton = document.getElementById('logout');
   logoutButton.addEventListener('click', logout);