+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v0.8.0
+
+- Add fixtures page to popup
+- Add statistics page with filtering and sorting to popup
+- Improve back button popup
+
 ### v0.7.0
 
 - Add individual league tables to popup (shows top 50)